+++ bindings/c++/src/image_reader-c++.h
@@ -19,6 +19,9 @@
 #ifndef SAIL_IMAGE_READER_CPP_H
 #define SAIL_IMAGE_READER_CPP_H
 
+#include <cstddef>
+#include <string>
+
 #ifdef SAIL_BUILD
     #include "error.h"
     #include "export.h"
@@ -26,8 +29,6 @@
     #include <sail/error.h>
     #include <sail/export.h>
 #endif
-
-#include <string>
 
 namespace sail
 {
@@ -89,8 +90,8 @@ public:
     /*
      * An interface to sail_start_reading_mem(). See sail_start_reading_mem() for more.
      */
-    sail_error_t start_reading(const void *buffer, unsigned long buffer_length, const plugin_info &splugin_info);
-    sail_error_t start_reading(const void *buffer, unsigned long buffer_length, const plugin_info &splugin_info, const read_options &sread_options);
+    sail_error_t start_reading(const void *buffer, size_t buffer_length, const plugin_info &splugin_info);
+    sail_error_t start_reading(const void *buffer, size_t buffer_length, const plugin_info &splugin_info, const read_options &sread_options);
 
     /*
      * An interface to sail_start_reading_io(). See sail_start_reading_io() for more.
